requirements.txt
pip
pytest
flake8
pandas
matplotlib
python-dotenv
moviepy
numpy
opencv-python
pillow
SoccerNet
timm
torch
torchvision
tqdm
wandb
huggingface_hub[cli]
pyzipper
opencv-python-headless==4.10.0.84